XC-297: Use `spl_associated_token_account_interface` crate
