+++ Cargo.toml
@@ -61,6 +61,7 @@ serde_json = "1.0.141"
 serde_with = "3.14.0"
 solana-account = "3.0.0"
 solana-account-decoder-client-types = "3.0.0"
+spl-associated-token-account-interface = "2.0.0"
 solana-client = "3.0.0"
 solana-clock = "3.0.0"
 solana-commitment-config = "3.0.0"
