+++ integration_tests/Cargo.toml
@@ -31,6 +31,7 @@ sol_rpc_client = { path = "../libs/client" }
 sol_rpc_types = { path = "../libs/types" }
 solana-account = { workspace = true }
 solana-account-decoder-client-types = { workspace = true }
+spl-associated-token-account-interface = { workspace = true }
 solana-commitment-config = { workspace = true }
 solana-hash = { workspace = true }
 solana-instruction = { workspace = true }
